requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
black==25.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
frozenlist==1.8.0
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
propcache==0.4.1
python-dotenv==1.2.1
pytokens==0.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
urllib3==2.5.0
yarl==1.22.0
